% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-iptrie.R
\name{as_iptrie}
\alias{as_iptrie}
\title{Convert a data frame to an IP trie}
\usage{
as_iptrie(x, ...)
}
\arguments{
\item{x}{a data frame with at least two columns. Unless you pass
two string column names via \code{...} the first two columns
in the data frame will be used.}

\item{...}{either do not specify extra parameter or specify exactly
two length 1 character vectors (i.e. "strings") with
the first being the column name to use for the IP/CIDR and
the second being the column name for the data/value.}
}
\description{
Pass in a data frame with column 1 full of IP addresses or CIDRs and
a column 2 full of character values and this will convert it to an
IP trie. You can use the \code{...} to positionally identify the column
names you want (as strings) to select.
}
\examples{
xdf <- data.frame(a = "10.1.10.0/24", b = "HOME", stringsAsFactors = FALSE)
as_iptrie(xdf)
}
