% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iptree-package.R
\docType{package}
\name{iptree}
\alias{iptree}
\alias{iptree-package}
\title{...}
\description{
\itemize{
\item URL: \url{https://gitlab.com/hrbrmstr/iptree}
\item BugReports: \url{https://gitlab.com/hrbrmstr/iptree/issues}
}
}
\author{
Bob Rudis (bob@rud.is)
}
