% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iptrie-package.R
\docType{package}
\name{iptrie}
\alias{iptrie}
\alias{iptrie-package}
\title{Efficiently Store and Query 'IPv4' Internet Addresses with Associated Data}
\description{
Tries are great ways to store data that has obvious hierarchichal
properties, such as 'IPv4' addresses. Methods are provided to create 'IPv4'
tries and store, retrieve and delete 'IPv4' address keys with values.
Functions are based on the 'zmap' 'iptree' 'C' library.
}
\details{
\itemize{
\item URL: \url{https://gitlab.com/hrbrmstr/iptrie}
\item BugReports: \url{https://gitlab.com/hrbrmstr/iptrie/issues}
}
}
\author{
Bob Rudis (bob@rud.is)
}
