% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{iptrie_destroy}
\alias{iptrie_destroy}
\title{Destroy an IP trie}
\usage{
iptrie_destroy(trie)
}
\arguments{
\item{trie}{a trie created with \code{\link[=iptrie_create]{iptrie_create()}}}
}
\value{
\code{NULL} (invisibly)
}
\description{
Destroies an IP trie
}
\examples{
x <- iptrie_create()
iptrie_insert(x, "10.1.10.0/24", "HOME")
iptrie_ip_exists(x,"10.1.10.1/32")
iptrie_ip_exists(x,"10.1.11.1/32")
iptrie_lookup(x, "10.1.10.1/32")
}
