% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{print.iptrie}
\alias{print.iptrie}
\title{Print for iptrie}
\usage{
\method{print}{iptrie}(x)
}
\arguments{
\item{x}{an iptrie}

\item{...}{ignored}
}
\value{
\code{trie} (invisibly)
}
\description{
Print for iptrie
}
\keyword{internal}
