% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{iptrie_create}
\alias{iptrie_create}
\alias{is_iptrie}
\title{Create a new IPv4 Trie}
\usage{
iptrie_create()

is_iptrie(x)
}
\arguments{
\item{x}{an \code{iptrie}}
}
\value{
an external pointer to an IP trie

an \code{iptrie}
}
\description{
Creates a new IPv4 Trie
}
\examples{
x <- iptrie_create()
iptrie_insert(x, "10.1.10.0/24", "HOME")
iptrie_ip_in(x,"10.1.10.1/32")
iptrie_ip_in(x,"10.1.11.1/32")
iptrie_lookup(x, "10.1.10.1/32")
}
