% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{iptrie_lookup}
\alias{iptrie_lookup}
\alias{iptrie_ip_exists}
\title{Lookup a value for an IPv4 Address+CIDR combo into an IPv4 Trie or Test for Existence}
\usage{
iptrie_lookup(trie, ip)

iptrie_ip_exists(trie, ip)
}
\arguments{
\item{trie}{a trie created with \code{\link[=iptrie_create]{iptrie_create()}}}

\item{ip}{an IPv4 address with mask (will use \code{/32} if not provided)}
}
\value{
character vector or \code{NULL}
}
\description{
Looks up a value for or tests existence of an IPv4 Address+CIDR combo into an IPv4 Trie
}
\examples{
x <- iptrie_create()
iptrie_insert(x, "10.1.10.0/24", "HOME")
iptrie_ip_exists(x,"10.1.10.1/32")
iptrie_ip_exists(x,"10.1.11.1/32")
iptrie_lookup(x, "10.1.10.1/32")
}
